+++ packages/desktop-app/src/renderer/shared/action-approval/manage/models/index.ts
@@ -1,4 +1,3 @@
 export { ModelsRemoveProviderApproval } from "./ModelsRemoveProviderApproval";
 export { ModelsSetDefaultApproval } from "./ModelsSetDefaultApproval";
-export { ModelsSetPeripheralApproval } from "./ModelsSetPeripheralApproval";
 export { ModelsUpsertProviderApproval } from "./ModelsUpsertProviderApproval";